Pegasus

+++ project1/sentiment_analysis/test.py
@@ -244,7 +244,7 @@ def check_pegasos_single_update():
     ex_name = "Pegasos single update"
 
     feature_vector = np.array([1, 2])
-    label, theta, theta_0 = 1, np.array([-1, 1]), -1.5
+    label, theta, theta_0 = 1, np.array([-1, 1],dtype=float), -1.5
     L = 0.2
     eta = 0.1
     exp_res = (np.array([-0.88, 1.18]), -1.4)
